refactor: move validation functions
