+++ islands/WhatsappLinkGenerator.tsx
@@ -1,76 +1,26 @@
+//? Manage form state
 import { useState } from "preact/hooks";
+//? Styled components
 import { StyledInput } from "../components/UI/StyledInput.tsx";
 import { CountryPhoneCodeSelect } from "../components/tools/CountryPhoneCodeSelect.tsx";
 //? Types for typecasting
 import { validationStatus } from "../types/misc/validationStatus.ts";
 import { StyledButton } from "../components/UI/StyledButton.tsx";
+//? Validation functions
+import { patternValidation } from "../services/form-validation/patternValidation.ts";
+import { validateAreaCode } from "../services/form-validation/validateAreaCode.ts";
+import { validatePhoneNumber } from "../services/form-validation/validatePhoneNumber.ts";
 
+//? Default form values and validation
 const defaultPhoneNumber = {
   countryCode: "93",
   areaCode: "",
   phoneNumber: "",
 };
-
 const defaultFormValidation = {
   areaCode: validationStatus.Unchanged,
   phoneNumber: validationStatus.Unchanged,
 };
-
-//? Validates the form's area code input field
-const areaCodeValidation = (
-  areaCode: string,
-): validationStatus.Invalid | validationStatus.Valid => {
-  //? Instantiate Area Code as a string length and if it's a valid number
-  const areaCodeStringLength = areaCode.length;
-  const isAreaCodeAnInvalidNumber = isNaN(Number(areaCode));
-
-  if (
-    areaCodeStringLength > 4 ||
-    isAreaCodeAnInvalidNumber
-  ) {
-    return validationStatus.Invalid;
-  } else {
-    return validationStatus.Valid;
-  }
-};
-
-//? Validates the form's phone number input field
-const phoneNumberValidation = (
-  areaCode: string,
-): validationStatus.Invalid | validationStatus.Valid => {
-  //? Instantiate Area Code as a string length and if it's a valid number
-  const areaCodeStringLength = areaCode.length;
-  const isAreaCodeAnInvalidNumber = isNaN(Number(areaCode));
-
-  if (
-    areaCodeStringLength < 5 ||
-    areaCodeStringLength > 10 ||
-    isAreaCodeAnInvalidNumber
-  ) {
-    return validationStatus.Invalid;
-  } else {
-    return validationStatus.Valid;
-  }
-};
-
-//? Given a current value, initial value and a validation function, return
-//? what is the state of validation of the input once the respective field
-//? loses focus
-function validateInput(
-  value: string,
-  initialState: string,
-  pattern: (
-    valueToValidate: string,
-  ) => validationStatus.Invalid | validationStatus.Valid,
-): validationStatus {
-  //? If the field is at initial state, reset validation
-  if (value === initialState) {
-    return validationStatus.Unchanged;
-  } //? If the field is not empty, check if a validation RegEx pattern was provided
-  else {
-    return pattern(value);
-  }
-}
 
 export default function WhatsappLinkGenerator() {
   const [phoneNumber, setPhoneNumber] = useState(defaultPhoneNumber);
@@ -113,10 +63,10 @@ export default function WhatsappLinkGenerator() {
           }}
           validationReference={formValidationStatus.areaCode}
           validationFunction={(input) => {
-            const result = validateInput(
+            const result = patternValidation(
               input.toString(),
               defaultPhoneNumber.areaCode,
-              areaCodeValidation,
+              validateAreaCode,
             );
             updateValidation((currentValidation) => ({
               ...currentValidation,
@@ -141,10 +91,10 @@ export default function WhatsappLinkGenerator() {
           }}
           validationReference={formValidationStatus.phoneNumber}
           validationFunction={(input) => {
-            const result = validateInput(
+            const result = patternValidation(
               input.toString(),
               defaultPhoneNumber.phoneNumber,
-              phoneNumberValidation,
+              validatePhoneNumber,
             );
             updateValidation((currentValidation) => ({
               ...currentValidation,
@@ -152,7 +102,7 @@ export default function WhatsappLinkGenerator() {
             }));
             return result;
           }}
-          helpInformation="Must only contain numerical digits and between 1 to 4 characters long"
+          helpInformation="Must only contain numerical digits and between 5 to 10 characters long"
         />
         <StyledButton classes="m-4 self-center" text="Generate link" />
       </form>
